+++ R/spacy_initialize.R
@@ -9,6 +9,11 @@
 #'   if \code{TRUE}, list available spaCy installations and prompt the user 
 #'   for which to use. If another (e.g. \code{python_executable}) is set, then 
 #'   this value will always be treated as \code{FALSE}.
+#' @param source_bash_profile logical; if \code{TRUE}, source \code{~/.bash_profile} before trying
+#'   to find python executables with spaCy installed. Most likely necessary to set \code{TRUE}, 
+#'   if using anaconda python in Mac or Linux. Default is \code{NULL} 
+#'   which functions as \code{TRUE} for non-Windows system (e.g. Mac/Linux) and 
+#'   \code{FALSE} for Windows system.
 #' @param virtualenv set a path to the python virtual environment with spaCy installed
 #'   Example: \code{virtualenv = "~/myenv"}
 #' @param condaenv set a path to the anaconda virtual environment with spaCy installed
@@ -18,6 +23,7 @@
 spacy_initialize <- function(model = "en", 
                              python_executable = NULL,
                              ask = FALSE,
+                             source_bash_profile = NULL,
                              virtualenv = NULL,
                              condaenv = NULL) {
 
@@ -26,37 +32,39 @@ spacy_initialize <- function(model = "en",
         message("spaCy is already initialized")
         return(NULL)
     }
+    # set an option for source_bash_profile
+    if(is.null(source_bash_profile)) {
+        if(Sys.info()['sysname'] == "Windows"){
+            source_bash_profile <- FALSE
+        } else {
+            source_bash_profile <- TRUE
+        }
+    }
     # once python is initialized, you cannot change the python executables
     if(!is.null(options("python_initialized")$python_initialized)) {
         message("Python space is already attached.  If you want to swtich to a different Python, please restart R.")
     } 
-    # a user can specify only one
-    else if(sum(!is.null(c(python_executable, virtualenv, condaenv))) > 1) {
-        stop(paste("Too many python environments are specified, please select only one",
-                   "from python_executable, virtualenv, and condaenv"))
+    else {
+        set_spacy_python_option(python_executable, 
+                                virtualenv, 
+                                condaenv, 
+                                ask, 
+                                source_bash_profile, 
+                                model)
     }
-    # give warning when nothing is specified
-    else if (sum(!is.null(c(python_executable, virtualenv, condaenv))) == 0){
-        # def_python <- ifelse(Sys.info()['sysname'] == "Windows", 
-        #                      system("where python", intern = TRUE), 
-        #                      system("which python", intern = TRUE))
-        message("Finding a python executable with spacy installed...")
-        spacy_python <- find_spacy(model, ask = ask)
-        if(!is.null(spacy_python)){
-            reticulate::use_python(spacy_python, required = TRUE)
-        } else {
-            stop("spaCy or language model ", model, " is not installed in any of python executables.")
-        }
-    } 
-    else {# set the path with reticulte
-        if(!is.null(python_executable)) {
-            if(check_spacy_model(python_executable, model) != "OK"){
-                stop("spaCy or language model ", model, " is not installed in ", python_executable)
+    
+    if (!is.null(options("spacy_python_setting")$spacy_python_setting)) {
+        ####
+        type <- options("spacy_python_setting")$spacy_python_setting$type
+        py_path <- options("spacy_python_setting")$spacy_python_setting$py_path
+        if(type == "python_executable") {
+            if(check_spacy_model(py_path, model) != "OK"){
+                stop("spaCy or language model ", model, " is not installed in ", py_path)
             }
-            reticulate::use_python(python_executable, required = TRUE)
+            reticulate::use_python(py_path, required = TRUE)
         }
-        else if(!is.null(virtualenv)) reticulate::use_virtualenv(virtualenv, required = TRUE)
-        else if(!is.null(condaenv)) reticulate::use_condaenv(condaenv, required = TRUE)
+        else if(type == "virtualenv") reticulate::use_virtualenv(py_path, required = TRUE)
+        else if(type == "condaenv") reticulate::use_condaenv(py_path, required = TRUE)
     }
     options("python_initialized" = TRUE) # next line could cause non-recoverable error 
     spacyr_pyexec(pyfile = system.file("python", "spacyr_class.py",
@@ -104,18 +112,28 @@ spacy_finalize <- function() {
 #'   if \code{TRUE}, list available spaCy installations and prompt the user 
 #'   for which to use. If another (e.g. \code{python_executable}) is set, then 
 #'   this value will always be treated as \code{FALSE}.
+#' @param source_bash_profile logical; if \code{TRUE}, source \code{~/.bash_profile} before trying
+#'   to find python executables with spaCy installed. Most likely necessary to set \code{TRUE}, 
+#'   if using anaconda python in Mac or Linux. 
 #' @keywords internal
 #' @importFrom data.table data.table
-find_spacy <- function(model = "en", ask){
+find_spacy <- function(model = "en", ask, source_bash_profile){
     spacy_found <- `:=` <- NA
     spacy_python <- NULL
     options(warn = -1)
+    if(source_bash_profile == TRUE & Sys.info()['sysname'] == "Windows"){
+        message("the option, source_bash_profile == TRUE, will be ignored for windows system")
+    }
     py_execs <- if(Sys.info()['sysname'] == "Windows") {
         system2("where", "python", stdout = TRUE)
+    } else if(source_bash_profile == TRUE) {
+        c(system2('source', '~/.bash_profile; which -a python', stdout = TRUE),
+          system2('source', '~/.bash_profile; which -a python3', stdout = TRUE))
     } else {
         c(system2('which', '-a python', stdout = TRUE),
           system2('which', '-a python3', stdout = TRUE))
     }
+    py_execs <- unique(py_execs)
     options(warn = 0)
     
     if (length(py_execs) == 0 | grepl("not find", py_execs[1])[1]){
@@ -176,4 +194,53 @@ check_spacy_model <- function(py_exec, model) {
     })
     options(warn = 0)
     return(paste(sys_message, collapse = " "))
+}
+
+
+set_spacy_python_option <- function(python_executable = NULL, 
+                                    virtualenv = NULL, 
+                                    condaenv = NULL, 
+                                    ask = NULL, 
+                                    source_bash_profile = NULL,
+                                    model = NULL) {
+    # a user can specify only one
+    if(sum(!is.null(c(python_executable, virtualenv, condaenv))) > 1) {
+        stop(paste("Too many python environments are specified, please select only one",
+                   "from python_executable, virtualenv, and condaenv"))
+    }
+    # give warning when nothing is specified
+    else if (sum(!is.null(c(python_executable, virtualenv, condaenv))) == 1){
+        if(!is.null(python_executable)) {
+            if(check_spacy_model(python_executable, model) != "OK"){
+                stop("spaCy or language model ", model, " is not installed in ", python_executable)
+            }
+            options(spacy_python_setting = list(type = "python_executable",
+                                                py_path = python_executable))
+        }
+        else if(!is.null(virtualenv)) {
+            options(spacy_python_setting = list(type = "virtualenv",
+                                                py_path = virtualenv))
+            
+        }
+        else if(!is.null(condaenv)) {
+            options(spacy_python_setting = list(type = "condaenv",
+                                                py_path = condaenv))
+        }
+    }
+    else if (!is.null(options("spacy_python_setting")$spacy_python_setting)) {
+        message("python path is already set\nspacyr will use: ", 
+                options("spacy_python_setting")$spacy_python_setting$type, " = ",
+                options("spacy_python_setting")$spacy_python_setting$py_path)
+    }
+    else {
+        message("Finding a python executable with spacy installed...")
+        spacy_python <- find_spacy(model, ask = ask, source_bash_profile = source_bash_profile)
+        if(!is.null(spacy_python)){
+            options(spacy_python_setting = list(type = "python_executable",
+                                                py_path = spacy_python))
+        } else {
+            stop("spaCy or language model ", model, " is not installed in any of python executables.")
+        }
+    }
+    return(NULL)
 }